Display GitLab cross reference in project admin
What
----

Go to project admin » references. You should see `gitlab_commit` in a
new section.

How
---

As soon as the gitlab plugin is enabled, the cross reference is
available for every projects.

Why
---

GitLab cross references behave differently than other cross references
in Tuleap:
* They are not inserted in reference/reference_group tables.
* They do not share the same inheritance behavior between projects.
* They cannot be enabled/disabled, even by site administrators.

This has been done this way for two main reasons:
* We should not break existing custom cross references named
'gitlab_commit'. (IMHO this is questionnable. My rapid tests concludes
that project references override system references and we could have
treated the gitlab cross references like other system ones without
breaking the existing. To be investigating further if needed.)
* If we consider `gitlab_commit` as a regular system cross reference,
what do we put in reference.link? This link is supposed to be a GitLab
one, or maybe we should use a Tuleap url that redirects to GitLab?

I don't know yet if/when we should change this behavior. In the meantime
I propose to at least display the cross reference in project admin
section, to be a mimimum consistent with other cross references.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I15b6f9c1e19fc810786e3c5fa542845f5d4fce0a