request #19182: Close user session as soon as possible when processing CSP violation reports
In a situation where a lot of CSP reports needs to be sent when opening a page
(it should not happen but ¯\_(ツ)_/¯) not closing the session as soon as possible
can increase the time needed to process them.

No functional change expected, the /csp-violations endpoint now closes
the PHP session as soon as possible.

Change-Id: I34886f10e28d1b32486dc10340fadc531aa2c24b