Fix cypress shortcuts tests and remove delay on modal
Part of story #18401 Disable keyboard shortcuts when a modal is open

How to test:
- make tests_cypress should succeed
- No functional change on TLP modals

Note:
The warning modals were dispatching an event twice with v-on=click. They
already listened to the modal's "close" event so no functional change
should occur.

The shortcuts tests were broken because of two reasons:
1. There is a delay of 300ms after closing a modal, before the "closed"
event is actually sent. During this time, keyboard shortcuts are
disabled.

As we could not find a trace of the reasons why this delay had been
implemented and as visually it looks like it makes no difference, it has
been removed.

2. The Vue title components are subject to dynamic import, which means
that there is a delay between the creation of an item and the time where
its title is actually displayed.

Using more specific data-test selectors ensures Cypress does not find
the "First item" when we ask it to wait for the "Last item".

Change-Id: I28399bed9316d3e27cb3e669d515a956c5f3b14c