Update bot api token
Go to GitLab project that is integrated in Tuleap and that already
received some comment from Tuleap (to add links to artifacts).

Revoke the access token used by Tuleap.

Commit with a link to an artifact (tuleap-xxx).

=> No link is added as comment to the commit.
=> In the logs you have "An error occurred during automatically comment
commit"

Now create a new access token, copy the token and issue the following
REST API request:

```
PATCH /gitlab_repositories
{
  "update_bot_api_token": {
    "gitlab_bot_api_token" : "The new token",
    "full_url" : "https://gitlab.example.com/path-to/project",
    "gitlab_internal_id" : 123
  }
}
```

Where 123 is the id of your GitLab project, and "The new token" the one
you copied just before. Full url is given by corresponding entry in
GET projects/:id/gitlab_repositories

Commit with a link to an artifact (tuleap-xxx).
=> In the logs you have "Comment was successfully added"
=> Links are added as comment to the commit

Now try to insert an invalid access token. The REST rout should return
an error and the invalid token is not saved in the database.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I8f6d4adf6629c67f009fdf81c8a542d4a6964644