Refactor Jira import to prepare for multiple import
Conversion from Jira data structures relied on hardcoded ids.
That worked well for the import of a single issueType but as
soon as you want to import more than one, that no longer work.

This refactoring should not change anything functionaly speaking
but ensures that IDs in the tuleap XML format are unique.

This also remove the dependency on `StatusValuesTransformer`,
list values are passed explicitely to the FieldMapping in order
to ensure the uniquness.

Part of story #18859 import a whole Jira project in CLI

Change-Id: I8ecb8904d7777d569988396f411054b633a85959