Replace deprecated sniff UnusedPrivateElements by Psalm
Only the detection of unused properties has been enabled for now (but
this was most of the work the sniff was doing anyway).
Since this require a run on the full sources it is only played in the
nightly pipeline.

A few fixes have been made to allow Psalm to properly analyze the code.
The GitActions class has been excluded, some properties are seen as
non used even so they are. This class is in a state that can be
qualified of hopeless so this should be good enough ¯\_(ツ)_/¯.

Part of request #19183: Update PHP coding standard tooling

Change-Id: I0e5e595953bc4359041ff467ac03a037a75c09a3