Artifact link type _is_child cannot be unused anymore
This is part of story #18857 define any type of artifact as children

The artifact link type _is_child is always enable in projects and cannot
be unused.

I wanted to extract the method artifactLinkTypeCanBeUnused
in a dedicated and unit tested object. However, I don't find any good
name for such an object.

Change-Id: I321f88dca62cffc9fd0b5ae06baad70b551d7e21