Bot to automatically comment on a merge request
If you create or edit a merge request in gitlab, then a comment is added
to it by Tuleap, like for commits.

Note: currently there is too much comments on a given MR. Every time its
status is changed, then the comment is added. We should add a comment
only if the content is edited. It will be done in a dedicated changeset.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I96e8f5ba669345cc51e0781fe79617b8c72b1c09