Updates MR data if it was already saved in DB
Create/update a GitLab merge request that references one artifact (in
title or description).
On this artifact, there is a link to the MR and its title is displayed
in the xref.

Edit the merge request, update the title and remove any reference to the
artifact.
Refresh the artifact, the xref displays the new title of the MR.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: Iecfc0ec89b28c02dd08aa199996010042f6b923f