Do not enforce _is_child link type if project uses link types
This is part of story #18857 define any type of artifact as children

Automatically set _is_child was kept for project that still not
use the artifact link types. In this case, hierarchy defines the
behaviour and this information must be set so that when this project
will switch to use artifact link types parent/child links will work as before.

Old mechanism of automatically reverse a link if there is a hierarchy
have also been removed.

Change-Id: If8b6cf3bcc86212d15161e2684897838b549fcad