REST route to regenerate webhook secret
You can issue the following REST call:

PATCH /gitlab_repositories
{
  "generate_new_secret": {
    "gitlab_repository_id": 2,
    "gitlab_repository_url": "https://gitlab.example.com/repo/url"
  }
}

This will:

1. Delete the previous webhook (if exists) in corresponding GitLab
project
2. Generate a new webhook secret for the repository
3. Create a new webhook in corresponding GitLab project
5. Save the id of the new webhook in DB

If we don't know the id of the old webhook (which is the case on
existing usages), then it will be kept in the GitLab project.
Maintainers will have to remove it by hand.

Note: you may need to adjust the SSL verification settings of the new
webhook depending on your environment.

Note: initial thoughts was to [edit the existing webhook], but
implementation seemed to be a little bit more complex than simply remove
and create.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

[edit the existing webhook] https://docs.gitlab.com/ee/api/projects.html#edit-project-hook

Change-Id: I7058d600cbd5a5d62aba86eb9f3be3ea6392071c