Store GitLab webhook id
The id of the GitLab webhook we create is now stored in the database. It
will be used later to be able to remove it (dedicated patch).

This changes the sequence of webhook creation. Before:

1. Generate a new webhook secret for the repository
2. Save it in DB
3. Create a new webhook in corresponding GitLab project

Now, the secret is saved alongside the webhook id:

1. Generate a new webhook secret for the repository
2. Create a new webhook in corresponding GitLab project and get its id
3. Save id+secret in DB

This means that if we are not able to create the webhook (GitLab server
error), then the secret is not anymore saved.

This change is transparent for the user. You can assert that the webhook
id is properly saved by comparing the id of the webhook on GitLab
project, and the id we saved for the repository.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I55027350807d27cdde79ab5cd4bfc5a9c24f788c