request #19212: Warn users of very old versions of Firefox and Chrome
A modal is now displayed to the users of very old versions of Firefox and
Chrome.

Versions older than Firefox 78 or Chrome/Chromium 84 are considered "very old".
It does NOT mean that anything more recent is supported, it only means it is
less likely to break completely when rendering a page of Tuleap.
As a reminder, only the latest version of these browsers are supported.

We should warn the users for multiple reasons:
* it is not a good thing for their security to use an unsupported critical
  software that browsers are nowadays so we should nudge them (and/or their
  IT department) in the right direction
* it can be a source of frustration for them because they might encounter
  unexpected bugs in Tuleap 🐛
* it can generate an unnecessary consumption of resources 🌱. For example,
  users of old versions of Chrome are able to use the new list picker but
  does not have the support for the lazy loading of images, they can end up
  loading a lot of user avatars they do not need
* it can negatively affect the performance of up to date browsers since under
  some circumstances we (Tuleap developers) can be forced to ship additional
  code and to polyfill features

Change-Id: If5def0293e60b8a09125a75cca3839fe71f111a0