Extract sending of a comment to a dedicated object
This is a refactoring to be able later to send an email when the
credentials are expired.

The new object does not do much, but will catch credentials error in a
dedicated changeset.

This is not done in postUrl because we don't want to send an email when
the user fails to create a repository. Only posting comment is
concerned.

No functional changes.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I7c0bd13e3b43703aa5f5bf7c4c747523834a9e31