Display "API token (personal or project)" instead of "API project token" in UI
This is part of story #17182: reference tuleap artifacts in gitlab merge requests

No functional changes.
You can still add a gitlab repository.
If you commit, a comment will be added on the commit automatically.

Now when we integrate a GitLab repository, we can use a project or a personal token,
 instead of juste a project token.
This token will be used to configure project hook in gitlab, and to add comments on commit and merge request.

Change-Id: I6c0c75127905ba256c06d06cf2d8043e993ae35e