request #19219: Creating a new tracker from a Jira import might not be available even when the instance can do background processing
The Jira import feature when you create a tracker is only available on
Tuleap instance that can do background processing.

However when the instance can do background processing but has not explicitly
set the configuration parameter sys_nb_backend_workers the feature is not be
available (it should). The issue is not unique to the Jira import, a similar
issue exist when processing pull request events, they might get processed
online when they could processed without blocking the end users.

This contribution exposes an new high level API to determine if the
instance is expected to be able to do some background processing.

The setting sys_async_emails has been removed. It was initially
introduced to allow a progressive deployment of the feature. The
background processing is now considered robust enough to always be
actived for all projects.

Change-Id: I1dbe098db28337e036bb6e9e5f4b474d3dc4beb1