request #19220: Force successive assignment statements to align their assignment tokens in PHP code
This rule is more or less done manually by main Tuleap contributors.
For consistency's sake, this rule is now enforced (and can be autofixed)
thanks to phpcs.

The change has been vastly autogenerated with phpcbf. Only a few files
were not automatically fixed because the assignment statement was
splitted on 2 lines.

The easiest way to "review" this is probably with 'git diff --ignore-space-change'.

Change-Id: I380be6cffe2d03766eff4297e3c72d74c700fbf7