Extract lib gettext
Part of story #18336 add support of markdown

No functional change expected in list-picker.
This lib will be needed for CKEditor-related code in Trackers. Also
list-picker has no more "hidden" dependencies on outside code, so it's
cleaner.

Note:
- Other gettext-related code such as "gettext-init" and
"gettext-factory" should be moved inside this library, but the change is
pretty big (a lot of apps depend on it) so it can be done later.

Change-Id: I14f2882d7123fa4075af227d7757f93bb7fde62a