Create new program increment
part of story #18408: plan Features in Program Increment

Given your user has permission to create a program,
then you should have a button "+Add a PI"

Once you have submitted the artifact you are redirected to the scaled agile plugin.

Please note that for now, there is no mention in the artifact view,that we are creating something specific to scaled plugin. it will be added in a dedicated commit

Change-Id: Ib8e3fa6a5000601e6a4293e8586aeebc7f6c9a27