Extract another lib for ckeditor image upload
Part of story #18336 add support of markdown

How to test:
- No functional change in "Image drag/drop" feature in the Tracker
Artifact view. It does not affect the Artifact modal.

Notes:
In order to provide a cleaner lib for CKEditor with markdown, we must
first extract this part.
It is very tied to the Artifact view form and how it functions, so in
the future it may be moved to Tracker plugin instead of Tuleap core, but
for now it must stay in core as long as RichTextEditor.js depends on it.

Some code has been changed to use classes instead of top-level functions
in order to leverage @tuleap/gettext lib (as libs can't do dynamic
import).
As a result of extracting this, there is no longer a need to add
regenerator/runtime in @tuleap/core Jest tests. It is contained in this
library now.
The lib has not been converted to TypeScript yet because typing became
too painful with CKEditor types. It should be done in a later
contribution.

Change-Id: I57f3f8c6299d1f256b326f4bb826b816d0cf5105