Block 'Add to the top backlog' action of the Agile Dashboard when in a program
Programs needs to have their own "top backlog" since they do not have
the same rule than the Agile Dashboard.

You might notice the workflow post action is still visible in the
administration. The action will be hidden in a upcoming contribution, as
it requires to change the way the tracker plugin does this detection.

Part of story #18408: plan Features in Program Increment

Change-Id: I88c935f137a6a733968ab59c97136b69d103cf4d