request #19199 User expiration date is randomly displayed
User expiration date has the correct value if we open the html,
thus value seems to be well set on db side.

However, the value displayed to the site admin is either erroneous or empty.
It seems to occur when the user language is set to French. The date is
not well formated for the datepicker so it cannot be well displayed
when using french language.

Change-Id: I26dc90385d182fe7a3b8161f3cf85a81cb1a0a86