Be able to navigate in list of tests in TestManagement with arrow keys
Part of story #18860 : Navigate with the keyboard in TestManagement

The list of tests has been implemented into a `tablist` to enhance its accessibility and usability.
It is considered one single element in which the user can choose and select a test using the arrow keys (similarly to a list of <option> in a <select>).
It makes it easier and quicker to tab and reach the test execution on the right side of the screen.

When a test execution is open, its test in the list will be the one taking focus.
If no test is open, the first one in the list will take focus.

.ts files have been added to the list of files taken into account when generating po`.

Change-Id: I489b03d1d58f1ffbbe51e992cceaefde4f1837c3