Add Markdown support in new follow-ups
Part of story #18336 add support of markdown

How to test:
- Rebuild tracker
- Go to the Artifact view, edit an existing artifact
- When writing a new follow-up comment, you can choose the "Markdown"
format. When you save, the follow-up comment will be saved in
"commonmark" format.
- Image drag/drop and upload still functions as before in the new
follow-up editor.
- Switching to HTML from Text or Markdown should "render" the text
content in HTML (Markdown is converted to HTML)
- Switching from HTML to Text or Markdown should convert the HTML to
Markdown (to preserve formatting as much as possible).
- No other functional change expected to the new follow-up editor.

Notes:
This contribution introduces a new tracker/rich-text-editor library. It
is meant to replace completely codendi/RichTextEditor.js +
tuleap/textarea_rte.js. The new lib is written in vanilla TypeScript in
order to have as few dependencies as possible and be easily testable and
(more importantly) easy to change, which is not the case in the former
codendi/tuleap combination.
The library resides in Tracker and not in core because it is very
dependant on FlamingParrot and Tracker Artifact view form. As for now it
will be used only in Tracker, it makes sense to have it here. Using it
elsewhere will likely require a lot of work and is not in the scope of
this story for now, but it should be possible to isolate the library
from such dependencies.
Some options have been voluntarily left behind. There is only one set of
options for CKEditor (others are not used).
Some functionality will be added in later patches. For example, there is
a snippet of code for default values. This will be done in future
contributions (it is not needed for the follow-up comment).

Change-Id: I9abf3f3124501c05d45a36448dd427c6ddc6bb54