request #19235: Log the reason when rejecting an ID token with invalid claims
The new structure making it easier to see which branch is not covered
some tests have been fixed to test what they are supposed to test and a
test have been added to cover a branch that was missed.

On a side note, mutation testing would have helped to detect those
incorrect tests. It could be an interesting option to explore in the
future at least for sensitive code paths like this one.

Change-Id: I53f061c5320a50f3a3481aeee63159de38f5cde1