Import links types
In order to simplify the management of errors, the Jira
source is trusted. That means that all the link types that
are being exposed by Jira will be created Tuleap side if
the same name doesn't exist yet.

It will be created even if not used because it's easier
to create everything and to (potentially) clean-up after
if site admin doesn't want to expose un-used types.

Part of story #18859 import a whole Jira project in CLI

Change-Id: Iaeed8c47fb224c59d1f6233b2b8cf0a0af44a426