Add help block handling to the Image upload form lib
Part of story #18336 add support of markdown

How to test:
- rebuild tuleap core
- No functional change in Image drag/drop feature in the Tracker
Artifact view. It does not affect the Artifact modal.

Notes:
The HelpBlock creation is not used by RichTextEditor.js in order to
avoid making too many changes in it. It will be used by a later
contribution that aims to replace RichTextEditor. It is pushed in this
patch in order to reduce the size of the next contribution and ease
review.

Using a factory instead of top-level functions as before is just to
avoid doing the gettext init multiple times per RichTextEditor instance.

Change-Id: Ia4f6d8be143b353216a7b3ec9eebc28d979d133c