Extract a lib for BreadcrumbPrivacy Vue component
Part of request #19247 Extract an internal lib for BreadcrumbPrivacy vue
component

How to test:
- Rebuild tuleap (make post-checkout)
- No functional change expected. The Breadcrumb for project privacy
functions as before in Document, Git, ScaledAgile and the "Switch to"
modal.

Note:
An additional patch is needed for Baseline plugin.
The unit test no longer really runs createPopover() from "tlp" but a
"no-op" mock instead. It was getting far too painful to make it work for
very limited gain. As a consequence, the snapshots have changed
marginally (a x-placement attribute and a CSS style have disappeared).
If you know how to make it work, be my guest and submit a contribution
:).

Change-Id: Ic1f1bbb5a58dc8461f0c63bca555da73581968a0