Introduce Jira Boards management
Board are the base of Agile management in Jira API. We need
them to be able to get the sprints. At this stage it's not
clear whether, within a given project, 2 different boards
can have a different list of Sprints (I didn't find a way
to do that). Hence, the current implementation takes the
shortest path: get the first scrum board of a given project
so we can get the sprints.

The Client Wrapper was changed to no longer carry the
Jira Core rest base URL as the client should be used to
query `/agile/` end points too.

Part of story #19242 create Jira project with Agile configuration

Change-Id: I460216cb82531b8172bfaa2fa892f7d47ae1978d