Convert post new thread to burning parrot
Posting a new thread in forumml is now done via a modal. The legacy
Flaming Parrot page is removed.

Same behavior is kept (adding/removing CC/attachments).

Note: The legacy url is still used to post the message. Since it is not
accepting GET anymore, the user that might have bookmarked the page may
encounter an [apparently internal server error][0] instead of a "404 Not
found".  Therefore the legacy url is kept to redirect to the threads
list where the user can find the button to post a new thread. Once the
POST will use a new url, then we will be able to remove the redirection.

Part of request #18846: Switch ForumML to BurningParrot

[0]: in that case, the status code is 405 (method not allowed), which is
more of a "client error" (4xx) than a "server error" (5xx). However
FrontRouter handles it as a server error.

Change-Id: I4d179c1d3d5eccc7c535649809f7fcdeef87f326