@tuleap/tlp depends on @types/jquery
When building the integrality of Tuleap it is not really an issue since
the dependency is provided by @tuleap/core. However when building a
subset of Tuleap like it can happen when building external packages, the
dependency might not be provided at all which will break the build.

Part of request #19247: Extract an internal lib for BreadcrumbPrivacy vue
component

Change-Id: I7fc3373213fe8a73ee51d28ed072028c588f8ce4