Introduce logic for an XML representation of trackers
Historicaly we only generated XML for tracker in one place:
tracker::exportXml() methods and related ones (formelement,
reports, etc).

With JiraImport, we started to generate XML out of an external
source (Jira Core API) and the chosen path was to output XML more
or less directly.

As we are going to create tracker XML from a 3rd source
(Jira Agile API) it's time to think about avoiding a duplication
of code (and errors). Hence introducing a re-useable API that
on top of the XML: XMLTracker and XMLxxxx that will come next.

As there are tons of parameters, I choose not to go with
dependency injection in constructor (fat constructor anti pattern)
so it's the Builder pattern that was chosen. But in order to
keep as much as possible of the immutability the "noSetters"
provide, the `withXXX` methods returns a brand new instance of
object: those are immutables.

It might be possible that the manipulation of SimpleXmlElement
can be made immutable too but that's not the scope of this
refactoring so far.

Part of story #19242 create Jira project with Agile configuration

Change-Id: I42aef4f2e286623c39a63ff60469b48dd706e0bb