Switch to… component does not work on FlamingParrot pages
Since Webpack does not have the capability to build ES2015 module to
export the vue-breadcrumb-privacy module [0], the library cannot be tree
shaken when imported. As a result in src/scripts/switch-to/src/components/Body/Projects/ProjectLink.vue
we need a dependency that is not needed for the part of the imported
library.

Issue introduced by 807fd75874c4bdb08afe98e2e385af97507a0266.

Part of request #19247: Extract an internal lib for BreadcrumbPrivacy vue
component

[0] https://github.com/webpack/webpack/issues/2933

Change-Id: Ibc8357150ac7bfe12e887baa811e281d86198792