Fix email copy paste broken in FP
Part of story #18336 add support of markdown
Introduced by a0ed87fa219948f67ecaf3d86ad5b928c3c8bb31

How to test:
- Set up a tracker with "reply by mail" enabled. There is a checkbox in
General options in administration and you also need some semantics like
Description and Title
- Edit an artifact from this tracker. Below the "new follow-up comment"
textarea, there is a little paragraph with a link. The link should open
a modal.
- No functional change in other parts of Tracker.

Notes:
I took this opportunity to separate scripts between "FlamingParrot"
where there are often dependencies on jQuery and codendi, and
"BurningParrot" which depends on tlp. The two should not be mixed as it
can encourage depending on tlp in FlamingParrot, which will crash. Same
for depending on codendi in BurningParrot.

Change-Id: I2357ec8b39690af4efe7625e9d6f80ed1d298530