Refactoring: wait for ready state interactive
Part of story #18336 add support of markdown

No functional change expected in the Artifact view when adding a new
follow-up comment.

Notes:
Instead of executing all the code immediately when loaded, it is a bit
safer to
wait for the DOM to be ready. Previous attempts were tried with
DOMContentLoaded, but there is some jQuery code that depends on
RichTextEditor and that was executing before it, causing bugs. This
attempt uses "onReadyStatechange" which executes at the same time as the
jQuery ready() method.

Change-Id: I868881bc7b9ee88a9ba74c6258be8f3bf9767d9d