Refactoring: Move Interface before FormatSelectorBuilder
Part of story #18336 add support of markdown

No functional change expected in Artifact view when adding a new
follow-up comment.

Notes:
FormatSelectorBuilder is also part of the "Display for FlamingParrot domain".
We should isolate our domain (RichTextEditor) from it.
If we were to change that code and use lit-html or mustache or some
other rendering library, we would not manipulate HTMLOptionElements and
HTMLDivElements. Letting the interface insert the Format selector also
leaves us the possibility of a "no-op" interface for cases where the
Format selectbox is created another way.

Change-Id: I7415dc7ecf556158167a537bde8167eba7c6f355