Refactoring : Using `ng-model` rather than dedicated functions in tests list filter of TestManagement
Part of story #18461 : Have keyboard shortcuts in TestManagement

This  makes use of `ng-model` to replace the `toggleStatus` and `toggleAutomatedTests` functions, so that the checkboxes directly set the status.
It also allows the initial checked status of checkboxes to be set through the controller only, instead of also using the `checked` attribute.

No functional changes expected.

Change-Id: Ic5e0383b93f8116c090357d30610b08c571ab542