Update badge color according to MR/PR status
The status of the GitLab MR or Tuleap PR should be reflected by a
different badge color in the cross reference in an artifact:

* Open/Review: secondary outline
* Merged : success outline
* Closed/Abandonned : danger outline

Now additional badge presenters does not rely anymore on boolean
parameters. Usage of named constructor should be easier to read.

Part of story #17182: reference tuleap artifacts in gitlab merge requests

Change-Id: I7d810a2397be8175ad688aab0c491b4d5c5d758a