request #19270: Use the build environnement based on Nix to build the RPM packages
With this change the RPM packages can be built within the Nix based
environment. This means we do not need the image enalean/tuleap-buildrpms
anymore (yeah, one less thing to maintain). It is also a good thing for
the reproducibility of the packages since every tools and code are now
contained inside the repository.

Usages of Perl in the specfile have been replaced by GNU sed to avoid
pulling Perl in the build environment.

No functional changes are expected.

Change-Id: I75faf1a94aca82f6a4180ff0e3ce3871c77c5f89