Build the BreadcrumbPrivacy lib as a ES module
This allow the library to be tree-shaken.

Since Webpack is not capable of building ES2015 module for libraries [0],
it is replaced by Vite [1]. The only inconvenient of this switch is that
Vite does not have a build watch mode at the moment [2] so we simulate
it using nodemon [3].

Note that we still also deliver the lib as UMD because Jest needs it.

Part of request #19287: Make internal libs "tree-shakeable"

[0] https://github.com/webpack/webpack/issues/2933
[1] https://vitejs.dev
[2] https://github.com/vitejs/vite/issues/1434
[3] https://nodemon.io/

Change-Id: I8341202dec6124c9704e630cbf348d5e02bb6746