Error console in FRS should not be thrown anymore
request #19289 Console error are thrown in "FRS/Files" when "News" service is not enabled

Error console are displayed when the user open the "New release" page
and wen the "Release Name" field lost the focus. These error occurs
because we want to hide or update elements which does not exist anymore
since 653d796d2cf5eeeda903c027775dda5a27206211.

How to test:
 - Go to the add new release page
=> No error displayed in the console.
 - Add a name for the release and then click somewhere else.
=> No error displayed in the console.

You should still be able to submit the form.

These errors was detected thanks to e2e tests.

Change-Id: Ifbddc889b4b3579d4d8351e852abca3e668dd0d9