Introduce BurningParrot for ML thread
In order to not produce a too big review, the full conversion to
BurningParrot will be done in multiple commits.

To test, either uncomment the call to OneThreadController::getUrl() in
ThreadsPresenterBuilder so that you can access the new view when
browsing threads, or build the url by yourself.

Like before (code has just been copied/arranged/tested), we should still
be able to display:

* text/plain messages
* text/plain messages with (downloadable) attachments
* text/html messages
* text/html messages with (downloadable) attachments
* text/html messages with inline images

Things that will not be kept on purpose:
* Display of sender's email address (to avoid exposing too much
personal information)
* Display of replies subject (OP's subject is good enough)
* Automatic detection of quotations in text/plain message (lines
starting with `>` being converted to <blockquote>). Let's see if users
complain about the lack of this feature.

Things to be done in next commits:
* Display monospace font switch
* Reply to a message
* Remove commented call and use OneThreadController::getUrl() everywhere
so that we can delete legacy code

Part of request #18846: Switch ForumML to BurningParrot

Change-Id: I1f6776b6da85d06a1de812db7d9bf6967d8b00b6