Export timetracking configuration
This is part of story #19253 import worklog as timetracking

Each exported issueType as a Tuleap tracker will now have timetracking
enabled by default (if timetracking plugin is enabled on Tuleap side).

This export a default configuration:

* Enabled
* "Project members" as writers
* No selected ugroups for readers.

Sprint is not an issueType on Jira side, so it will not have timetracking enabled yet.

In addition, this patch fixes a bug on the Tuleap timetracking XML import feature.

Change-Id: I3a46bf3d5c614e0917b53e62efdd57a2e5379bb7