Extract edition of campaign label in dedicated modal
In TTM » Campaign, the edition of the label of the campaign is now done
in a dedicated modal. This will avoid to load 4000 tests just to rename
the campaign.

The configuration of automated tests will also be moved in its dedicated
modal, but it will be done in a dedicated patch to not clutter the
review. As such the label of the legacy [🖊️ edit] button is not changed yet.

Part of request #18498: Separate campaing edition modal in 3 distinct modals

Change-Id: I816cbc548a7e099e1af584e597fb21204313fe71