Introduce Jira platform configuration handling
This is part of story #19253 import worklog as timetracking

Jira tracked times can only be imported in Tuleap if "JIRA" is the
provider of the timetracking in the jira platform.

To check this, a new mechanism about platform configuration retrieving
is introduced in this patch. This mechanism is done only once before
retrieving Jira project data, and provided to objects that do the
import. Tuleap external plugins can also do their own platform check if
necessary (done here by timetracking plugin).

Times are not yet imported, but somethig is logged to know if times can
be imported based on the platform configuration (the same pattern that
was used to initiate Jira board import).

No unit tests about these logs have been added, because it will be
replaced by the real and unit tested time import feature in the next
commit.

Change-Id: Ibe602f7d68701c82a97d4a2442ff4d9d79eeef06