Extract configuration of automated tests in dedicated modal
In TTM » Campaign, the configuration of automated tests for the campaign
is now done in a dedicated modal. This will avoid to load 4000 tests
just to update this configuration.

Since the legacy modal is only dealing with tests selection, the Edit
button is renamed and get a new icon.

This should close request #18498: Separate campaing edition modal in 3 distinct modals

Change-Id: I96b86e5467c8190cfa7970918de3b6f245582e03