Enforce isolated modules
Part of request #19300 Enforce
@typescript-eslint/consistent-type-imports

Set isolatedModules in tsconfig. This will warn people when exporting
types.

No functional change expected.

Note that we have to sometimes add an empty export {} so that some files
without any import or export will be considered as modules as well.
I also removed the commented lines in our main tsconfig file. IDEs will
tell you an accurate information when hovering properties and the format
can be found online and up-to-date.

Change-Id: Ic5e176a309dbfff3ef023e951b505ef924899a29