request #19301 TTM : errors are not displayed when the status update fail
When something prevents to switch the status of a test execution,
nothing is displayed in the TTM interface. One has to open the browser
console to know that there was an error. This can be reproduced by
setting a field dependencies between a random field and a the status
field.

For example with a source field F with values [A, B, C, D] and the
following dependencies:
A --> notrun
B --> passed
C --> failed
D --> blocked

With this patch, if the update of the test execution is not possible,
the user should be warned.

Change-Id: I3bb0a5b09b24e7f53e76effcbd960a59d8c55ebd