Rename `Show non-passed tests` to `Show remaining tests` in TTM
Part of story #18461 : Have keyboard shortcuts in Test Management

So it does not collide with `n` `New campaign` in Campaigns list page,
we change the shortcut for `Show non-passed tests`.
We also rename this shortcut to `Show remaining tests`. This allow to
use an obvious shortcut key `r` and makes it more elegant in both English
and French.

Change-Id: Id87dfde3b4bce4d71595f93d4817556468170fe1